Include in Footer
Included in the Footer of our library's EBSCO Linksource page, this
script takes text and links out of the default table view and marks them
up in styleable divs and paragraphs.  It's very individual to our
library: each #3863_Section (etc.) section is defined at the level of
each A-to-Z/LinkSource site.